requirements.txt
# Core dependencies
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
serpapi>=0.1.5

# AI/NLP
anthropic>=0.18.0
tiktoken>=0.5.0
textblob>=0.17.1
nltk>=3.8.0

# Database
sqlite-utils>=3.35.0

# CLI & utilities
click>=8.1.0
rich>=13.5.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Export formats
openpyxl>=3.1.0
python-docx>=1.0.0

# Async support
asyncio-throttle>=1.0.0
